Removed extra xml files Created central config file, evebot.xml, currently only stores ORE info. Created obj_Asteroid Created obj_Ship created obj_Drones created obj_EVEBotUI created obj_Cargo created obj_Station
All function calls which could be converted to methods or members have been, for those functions.

Note that very little changes have occured in oCombat, except where necessary for it to call a function which was moved.

Rewrote cargo transfer routines, they should be 100% reliable now, and extensible for other cargo types in the future.

Created basic use of bookmarked asteroid locations, any bookmarks that start with Belt: or Belt will be iterated thru at random.  This is not up to par yet, I wouldn't use it.  Need to integrate bookmark iteration with blacklisting for empty (or empty of desired) ore types and fall back to overview list of belts if so (which should also be added to blacklist).

Rewrote EB_GlobalVariablesDaemon.iss, see obj_EVEMONUI.iss for an example of how to do this without making a 2nd script.